requirements.txt
sqlalchemy
psycopg2-binary
fastapi
uvicorn[standard]
python-dotenv
twilio
python-multipart
cryptography
Pillow
fpdf2
requests
openai
anthropic
httpx
